Patch ICU build scripts for broken mkinstalldirs on Windows.
The ancient mkinstalldirs script attempts to make directories called "C:"
on Windows under MSYS2. This succeeds, but results in a directory that
violates the Windows naming policies and can't easily be deleted, causing
weird errors if you try to rm -r (or similar) the ICU build directory.

Best as I can tell, mkinstalldirs as it exists in the ICU source tree is
mkdir -p, so we patch the ICU source to just use mkdir -p instead of the
script on Windows/MinGW.