Perform ICU building in a subdir of the project build folder.
You can encounter weird issues if we try to share accidentally share the
same build directory across e.g. GCC and Clang on the same machine.
This setup should allow you to share the same source archive (and not
have to re-download every time you blow away your build dir) while
ensuring that every build directory has its own independent copy of the
library. In the event that the same build directory is used for a
project that bumps its required ICU version (which can happen if you've
got a dev switching between branches with different ICU versions), we
now add the version as a suffix to the folder names to avoid weird
conflicts and unnecessary rebuilds.